:bug: Small refactor
